requirements.txt
annotated-types==0.7.0
anyio==4.4.0
apturl==0.5.2
attrs==21.2.0
bcrypt==3.2.0
blinker==1.8.2
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
editdistpy==0.1.4
fasteners==0.14.1
filelock==3.14.0
Flask==3.0.3
future==0.18.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
joblib==0.17.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lingua-language-detector==2.0.2
lockfile==0.12.2
louis==3.20.0
lxml==5.2.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
nltk==3.7
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
openai==1.30.3
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.1
pluggy==0.13.0
polib==1.2.0
portalocker==2.8.2
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymystem3==0.2.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyspellchecker==0.8.1
pytest==6.2.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2021.11.10
reportlab==3.6.8
requests==2.25.1
sacrebleu==2.4.2
SecretStorage==3.3.1
setuptools==59.6.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.1
symspellpy==6.7.7
systemd-python==234
tabulate==0.9.0
toml==0.10.2
tqdm==4.57.0
typing_extensions==4.12.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.26.1
wadllib==1.3.6
Werkzeug==3.0.3
wheel==0.37.1
xdg==5
xkit==0.0.0
zipp==1.0.0
